requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Email Processing
imaplib2==3.6
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.2

# Database (Optional - for production)
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Security
cryptography==41.0.8
bcrypt==4.1.2
PyJWT==2.8.0

# Logging and Monitoring
loguru==0.7.2
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Development Tools
ipython==8.18.1
jupyter==1.0.0
notebook==7.0.6

# File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
Pillow==10.1.0

# HTTP and API
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Machine Learning (Optional)
transformers==4.36.2
torch==2.1.2
tensorflow==2.15.0

# Performance
uvloop==0.19.0
orjson==3.9.10

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocstrings[python]==0.24.0 